added api

+++ OnionArchitectureDemo.ApplicationServices/Cars/Queries/GetCarList/CarModel.cs
@@ -2,7 +2,7 @@
 {
     public class CarModel
     {
-        public double SalesListingId { get; set; }
+        public long SalesListingId { get; set; }
         public string CarName { get; set; }
     }
 }
